Rename `thprac_main.*` to `attach.*`
