+++ thprac/src/thprac/attach.cpp
@@ -1,4 +1,4 @@
-﻿#include "thprac_main.h"
+﻿#include "attach.h"
 #include "thprac_games.h"
 #include "thprac_gui_locale.h"
 #include "thprac_launcher_wnd.h"
