+++ thprac/src/thprac/thprac_launcher_games.cpp
@@ -8,7 +8,7 @@
 #include "thprac_launcher_utils.h"
 #include "thprac_launcher_wnd.h"
 #include "thprac_load_exe.h"
-#include "thprac_main.h"
+#include "attach.h"
 #include "thprac_utils.h"
 #include <Windows.h>
 #include <shlwapi.h>
